+++ connectors/cache/fallback_test.go
@@ -1136,6 +1136,23 @@ func TestCacheableEntities(t *testing.T) {
 	assert.Len(t, set, 1)
 }
 
+// Test setting cacheable endpoints
+func TestCacheableEndpoints(t *testing.T) {
+	c := NewConnector(memory.NewConnector(), memory.NewConnector(), nil, nil)
+	assert.Len(t, c.cacheableEndpointStatus, 0)
+
+	endpoints := []string{"getEaterPromotions", "getPromotionsForStores"}
+	c = NewConnector(memory.NewConnector(), memory.NewConnector(), nil, nil, SetCacheableEndpoints(endpoints...))
+	assert.Len(t, c.cacheableEndpointStatus, 2)
+
+	// Test context key setters and getters
+	for _, endpoint := range endpoints {
+		ctx := context.Background()
+		ctx = SetContextEndpoint(ctx, endpoint)
+		assert.Equal(t, GetContextEndpoint(ctx), endpoint)
+	}
+}
+
 func TestWriteKeyValueToFallback(t *testing.T) {
 	connector := NewConnector(memory.NewConnector(), memory.NewConnector(), nil, nil)
 	err := connector.writeKeyValueToFallback(context.TODO(), testEi, "a", nil)